# storage/app/api-docs.json
{"openapi":"3.0.0","info":{"title":"Simple Auth API","contact":{"name":"Rizal Fauzie R.","email":"rizal@fauzie.my.id"},"version":"1.0"},"paths":{"\/api\/register":{"put":{"tags":["Authentication"],"summary":"User registration","description":"User Registration","operationId":"\/register","requestBody":{"description":"User details to registered","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/NewUser"}}}},"responses":{"201":{"description":"Registration Succeess","content":{"application\/json":{"schema":{"properties":{"status":{"description":"User registration","type":"boolean"},"code":{"description":"User registration","type":"number"}},"type":"object"}}}},"422":{"description":"Error: Parameter validation Failed.","content":{"application\/json":{}}},"404":{"description":"Error: Registration Failed.","content":{"application\/json":{}}}}}},"\/api\/login":{"post":{"tags":["Authentication"],"summary":"Authenticate user by email and password\nand return current user token to use on Authorization header.","description":"Authenticate User","operationId":"\/login","parameters":[{"name":"email","in":"query","description":"User email address","required":true,"schema":{"type":"string","minimum":6}},{"name":"password","in":"query","description":"User password","required":true,"schema":{"type":"string","minimum":3}}],"responses":{"200":{"description":"Return JSON object with token for API authentication.","content":{"application\/json":{"schema":{"properties":{"status":{"description":"Authenticate user by email and password\nand return current user token to use on Authorization header.","type":"boolean"},"token":{"description":"Authenticate user by email and password\nand return current user token to use on Authorization header.","type":"string"}},"type":"object"}}}},"401":{"description":"Error: Unauthorized.","content":{"application\/json":{"schema":{"properties":{"status":{"description":"Authenticate user by email and password\nand return current user token to use on Authorization header.","type":"boolean"}},"type":"object"}}}},"422":{"description":"Error: Parameter validation Failed.","content":{"application\/json":{}}}}}},"\/api\/logout":{"get":{"tags":["Logged In"],"summary":"Logout current user.","description":"User Logout","operationId":"\/logout","responses":{"200":{"description":"Success or failure user logged out from status property.","content":{"application\/json":{"schema":{"properties":{"status":{"description":"Logout current user.","type":"boolean"}},"type":"object"}}}}},"security":["Authorization"]}},"\/api\/exists":{"post":{"tags":["Utilities"],"summary":"Check if email exists.","description":"Check if email is exists","operationId":"\/exists","parameters":[{"name":"email","in":"query","description":"Valid email address","required":true,"schema":{"type":"string","minimum":6}}],"responses":{"200":{"description":"Email exists status from status property.","content":{"application\/json":{"schema":{"properties":{"status":{"description":"Check if email exists.","type":"boolean"}},"type":"object"}}}}}}},"\/api\/unique":{"post":{"tags":["Utilities"],"summary":"Check if email unique.","description":"Check if email is unique","operationId":"\/unique","parameters":[{"name":"email","in":"query","description":"Valid email address","required":true,"schema":{"type":"string","minimum":6}}],"responses":{"200":{"description":"Email unique status from status property.","content":{"application\/json":{"schema":{"properties":{"status":{"description":"Check if email unique.","type":"boolean"}},"type":"object"}}}}}}},"\/api\/user":{"get":{"tags":["Logged In"],"summary":"Get current User Details.","description":"Get User Details","operationId":"\/user","responses":{"200":{"description":"Success fetching user data object.","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/User"}}}},"401":{"description":"Error: Unauthorized."}},"security":["Authorization"]}}},"components":{"schemas":{"NewUser":{"required":["firstname","email","phone","gender","password"]},"UserLogin":{"required":["email","password"]},"User":{"type":"object","allOf":[{"$ref":"#\/components\/schemas\/NewUser"},{"properties":{"id":{"type":"integer","format":"int64"},"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"dob":{"type":"string"},"gender":{"type":"string"},"active":{"type":"boolean"},"password":{"type":"string","format":"password"},"api_token":{"type":"string","format":"password"},"remember_token":{"type":"string","format":"password"},"created_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}},"type":"object"}]}},"securitySchemes":{"Authorization":{"type":"http","description":"API Access Authorization","name":"Authorization","scheme":"bearer"}}}}